Move package handler content to consumer package
