+++ consumer/consumer_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/heetch/felice/consumer"
-	"github.com/heetch/felice/consumer/handler"
 	"github.com/heetch/felice/message"
 )
 
@@ -12,7 +11,7 @@ import (
 func TestHandleLogs(t *testing.T) {
 	tl := consumer.NewTestLogger(t)
 	c := &consumer.Consumer{Logger: tl.Logger}
-	c.Handle("foo", handler.HandlerFunc(func(m *message.Message) error {
+	c.Handle("foo", HandlerFunc(func(m *message.Message) error {
 		return nil
 	}))
 	tl.LogLineMatches(`Registered handler. topic="foo"`)
